projects/capstone/starter/requirements.txt
alembic==1.6.5
asn1crypto==0.24.0
attrs==21.2.0
cffi==1.14.6
click==8.0.1
cryptography==2.6.1
entrypoints==0.3
ez-setup==0.9
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.0.4
importlib-metadata==4.6.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==17.1.1
keyrings.alt==3.1.1
Mako==1.1.4
MarkupSafe==2.0.1
packaging==20.9
pkg-resources==0.0.0
pluggy==0.13.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
SecretStorage==2.3.1
six==1.12.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
